fix packaging due to wrong names and wrong path
