--- src/main/java/com/develhope/spring/vehicle/ExcessiveParameterException.java
@@ -1,9 +1,0 @@
-package com.develhope.spring.vehicle;
-
-
-public class ExcessiveParameterException extends Exception{
-
-    public ExcessiveParameterException(String message) {
-        super(message);
-    }
-}
