--- src/main/java/com/develhope/spring/vehicle/optionals/enums/TypeOfStereos.java
@@ -1,10 +1,0 @@
-package com.develhope.spring.vehicle.optionals.enums;
-
-public enum TypeOfStereos {
-    FULLOPTIONAL,
-    BASE,
-    INFOTAINMENT,
-    CDREADER,
-    BLUETOOTH,
-    AUX;
-}
