--- src/main/java/com/develhope/spring/vehicle/optionals/models/Optionals.java
@@ -1,9 +1,0 @@
-package com.develhope.spring.vehicle.optionals.models;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.Table;
-
-
-public interface Optionals {
-
-}
