--- src/main/java/com/develhope/spring/vehicle/repositories/VehicleRepository.java
@@ -1,9 +1,0 @@
-package com.develhope.spring.vehicle.repositories;
-
-import com.develhope.spring.vehicle.models.Vehicle;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface VehicleRepository extends JpaRepository<Vehicle, Long> {
-}
