--- src/main/java/com/develhope/spring/vehicle/vehicleEnums/Colors.java
@@ -1,10 +1,0 @@
-package com.develhope.spring.vehicle.vehicleEnums;
-
-public enum Colors {
-    RED,
-    BLUE,
-    YELLOW,
-    BLACK,
-    WHITE,
-    GREY
-}
