--- src/main/java/com/develhope/spring/vehicle/vehicleEnums/Gears.java
@@ -1,6 +1,0 @@
-package com.develhope.spring.vehicle.vehicleEnums;
-
-public enum Gears {
-    MANUAL,
-    AUTOMATIC
-}
