--- src/main/java/com/develhope/spring/vehicle/vehicleEnums/MarketStatus.java
@@ -1,7 +1,0 @@
-package com.develhope.spring.vehicle.vehicleEnums;
-
-public enum MarketStatus {
-    ORDERABLE,
-    AVIABLE,
-    NOTAVIABLE
-}
