--- src/main/java/com/develhope/spring/vehicle/vehicleEnums/MotorPowerSupply.java
@@ -1,10 +1,0 @@
-package com.develhope.spring.vehicle.vehicleEnums;
-
-public enum MotorPowerSupply {
-    DIESEL,
-    GASOLINE,
-    METHANE,
-    GPL,
-    ELECTRIC,
-    HYBRID
-}
