+++ src/main/java/com/develhope/spring/vehicles/models/ExcessiveParameterException.java
@@ -1,0 +1,9 @@
+package com.develhope.spring.vehicles.models;
+
+
+public class ExcessiveParameterException extends Exception{
+
+    public ExcessiveParameterException(String message) {
+        super(message);
+    }
+}
