+++ src/main/java/com/develhope/spring/vehicles/optionals/enums/TypeOfStereos.java
@@ -1,0 +1,10 @@
+package com.develhope.spring.vehicles.optionals.enums;
+
+public enum TypeOfStereos {
+    FULLOPTIONAL,
+    BASE,
+    INFOTAINMENT,
+    CDREADER,
+    BLUETOOTH,
+    AUX;
+}
