+++ src/main/java/com/develhope/spring/vehicles/optionals/models/Optionals.java
@@ -1,0 +1,9 @@
+package com.develhope.spring.vehicles.optionals.models;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.Table;
+
+
+public interface Optionals {
+
+}
