+++ src/main/java/com/develhope/spring/vehicles/repositories/VehicleRepository.java
@@ -1,0 +1,9 @@
+package com.develhope.spring.vehicles.repositories;
+
+import com.develhope.spring.vehicles.models.Vehicle;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface VehicleRepository extends JpaRepository<Vehicle, Long> {
+}
