+++ src/main/java/com/develhope/spring/vehicles/vehicleEnums/Colors.java
@@ -1,0 +1,10 @@
+package com.develhope.spring.vehicles.vehicleEnums;
+
+public enum Colors {
+    RED,
+    BLUE,
+    YELLOW,
+    BLACK,
+    WHITE,
+    GREY
+}
