+++ src/main/java/com/develhope/spring/vehicles/vehicleEnums/Gears.java
@@ -1,0 +1,6 @@
+package com.develhope.spring.vehicles.vehicleEnums;
+
+public enum Gears {
+    MANUAL,
+    AUTOMATIC
+}
