+++ src/main/java/com/develhope/spring/vehicles/vehicleEnums/MarketStatus.java
@@ -1,0 +1,7 @@
+package com.develhope.spring.vehicles.vehicleEnums;
+
+public enum MarketStatus {
+    ORDERABLE,
+    AVIABLE,
+    NOTAVIABLE
+}
