+++ src/main/java/com/develhope/spring/vehicles/vehicleEnums/MotorPowerSupply.java
@@ -1,0 +1,10 @@
+package com.develhope.spring.vehicles.vehicleEnums;
+
+public enum MotorPowerSupply {
+    DIESEL,
+    GASOLINE,
+    METHANE,
+    GPL,
+    ELECTRIC,
+    HYBRID
+}
