+++ src/test/java/com/develhope/spring/vehicles/services/VehicleServiceTest.java
@@ -1,0 +1,4 @@
+package com.develhope.spring.vehicles.services;
+
+public class VehicleServiceTest {
+}
